Refactor package
* Create _helpers.py
* Create _parts.py
* Create _parsing.py
* Create _expressions.py
* Remove _definitions.py
